Move pusher connection stuff to its own file
